Use only one instance of the adaptor
Handlers are registered to a singleton instance of MathJax.
Since the creation of the renderer has been moved from the attacher to the transformer, the previous implementation would register a new HTML handler for each document processed and never unregister it.
All documents may use the same adaptor (and registered handler) because no configuration is currently supported for the handlers.