Move adjustement of options format to the attacher
The code that would remove the `tex` options from the output options was duplicated in the `output-chtml` and `output-svg` files.
It was moved to the attacher instead, and the signature of the `createRenderer` function was adjusted to reflect the separation of the input and output options.
MathJax does not support `tex: undefined` in the output options, so `delete` has to be used.

The options for the browser plugin correspond to the TeX input processor.
This would conflict with the previous hack used for the non-browser plugin.
Hence, a `browser` flag was added to the `createPlugin` function.