Change to replace wrapping elements
Previously, spans and divs were generated by `remark-math`.
They were left as-is by `rehype-katex` and `rehype-mathjax`.
With fc32531, (`<pre>` and) `<code>` elements are generated by
`remark-math`.
That is to allow folks to use normal markdown, as in ` ```math `,
to generate math blocks.
However, `<span>` and `<div>` do not contribute to how a document
is displayed by browsers, but `<pre>` and `<code>` do.
To solve that, this commit *replaces* those elements instead of
changing their contents, when using `rehype-katex` or
`rehype-mathjax`.